+++ src/async_adapters/mod.rs
@@ -3,5 +3,8 @@
 #[cfg(feature = "tokio")]
 pub mod tokio_adapter;
 
+#[cfg(feature = "tokio")]
+pub mod tokio_injector;
+
 #[cfg(feature = "channel")]
 pub mod channel;